refactor: cleanup old wallet import/create code
