+++ src/redux/slices/settingsSlice/index.ts
@@ -4,8 +4,10 @@ import { ColorValue, Platform } from 'react-native'
 import { initializeSslPinning } from 'react-native-ssl-public-key-pinning'
 import { RifWalletServicesFetcher } from '@rsksmart/rif-wallet-services'
 import { providers } from 'ethers'
+import { RifRelayConfig } from '@rsksmart/rif-relay-light-sdk'
 
-import { deleteCache, getRifRelayConfig } from 'core/operations'
+import { ChainID, WalletState } from 'lib/eoaWallet'
+
 import { deleteDomains } from 'storage/DomainsStore'
 import { deleteContacts as deleteContactsFromRedux } from 'store/slices/contactsSlice'
 import { resetMainStorage } from 'storage/MainStorage'
@@ -39,8 +41,8 @@ import {
 } from 'store/slices/persistentDataSlice'
 import { Wallet } from 'shared/wallet'
 import { addressToUse } from 'shared/hooks'
-import { ChainID } from 'src/lib/eoaWallet'
 import { createAppWallet, loadAppWallet } from 'src/shared/utils'
+import { MMKVStorage } from 'storage/MMKVStorage'
 
 import {
   Bitcoin,
@@ -51,6 +53,32 @@ import {
 } from './types'
 import { UsdPricesState } from '../usdPricesSlice'
 import { BalanceState } from '../balancesSlice/types'
+
+export const deleteCache = () => {
+  const cache = new MMKVStorage('txs')
+  cache.deleteAll()
+}
+
+export const getRifRelayConfig = (chainId: 30 | 31): RifRelayConfig => {
+  return {
+    smartWalletFactoryAddress: getWalletSetting(
+      SETTINGS.SMART_WALLET_FACTORY_ADDRESS,
+      chainTypesById[chainId],
+    ),
+    relayVerifierAddress: getWalletSetting(
+      SETTINGS.RELAY_VERIFIER_ADDRESS,
+      chainTypesById[chainId],
+    ),
+    deployVerifierAddress: getWalletSetting(
+      SETTINGS.DEPLOY_VERIFIER_ADDRESS,
+      chainTypesById[chainId],
+    ),
+    relayServer: getWalletSetting(
+      SETTINGS.RIF_RELAY_SERVER,
+      chainTypesById[chainId],
+    ),
+  }
+}
 
 const sslPinning = async (chainId: ChainTypesByIdType) => {
   const rifWalletServiceDomain = getWalletSetting(
@@ -203,7 +231,7 @@ export const createWallet = createAsyncThunk<
 })
 
 export const unlockApp = createAsyncThunk<
-  void,
+  WalletState,
   UnlockAppAction,
   AsyncThunkWithTypes
 >('settings/unlockApp', async (payload, thunkAPI) => {
@@ -278,8 +306,6 @@ export const unlockApp = createAsyncThunk<
       return thunkAPI.rejectWithValue('No Existing Wallet')
     }
 
-    // const { kms, rifWallet, rifWalletIsDeployed } = existingWallet
-
     // set wallet and walletIsDeployed in WalletContext
     initializeWallet(wallet, {
       isDeployed: await wallet.isDeployed,
@@ -326,39 +352,6 @@ export const resetApp = createAsyncThunk(
     }
   },
 )
-
-// Not currently used, we'll be needed when support
-// for multiple wallets
-// export const addNewWallet = createAsyncThunk(
-//   'settings/addNewWallet',
-//   async ({ networkId }: AddNewWalletAction, thunkAPI) => {
-//     try {
-//       const keys = await getKeys()
-//       const { chainId } = thunkAPI.getState().settings
-//       if (!keys) {
-//         return thunkAPI.rejectWithValue(
-//           'Can not add new wallet because no KMS created.',
-//         )
-//       }
-//       const { kms } = KeyManagementSystem.fromSerialized(keys)
-//       const { rifWallet, isDeloyed } = await addNextWallet(
-//         kms,
-//         createRIFWalletFactory(
-//           request => thunkAPI.dispatch(onRequest({ request })),
-//           chainId,
-//         ),
-//         networkId,
-//       )
-//       thunkAPI.dispatch(setNewWallet({ rifWallet, isDeployed: isDeloyed }))
-//       return {
-//         rifWallet,
-//         isDeloyed,
-//       }
-//     } catch (err) {
-//       return thunkAPI.rejectWithValue(err)
-//     }
-//   },
-// )
 
 const initialState: SettingsSlice = {
   isSetup: false,
@@ -410,11 +403,6 @@ const settingsSlice = createSlice({
     setPreviouslyUnlocked: (state, { payload }: PayloadAction<boolean>) => {
       state.previouslyUnlocked = payload
     },
-    // Not currently used, we'll be needed when support
-    // for multiple wallets
-    // switchSelectedWallet: (state, { payload }: PayloadAction<string>) => {
-    //   state.selectedWallet = payload
-    // },
     resetKeysAndPin: () => {
       deleteKeys()
       deleteDomains()
@@ -456,15 +444,6 @@ const settingsSlice = createSlice({
     builder.addCase(unlockApp.fulfilled, state => {
       state.loading = false
     })
-    // builder.addCase(addNewWallet.pending, state => {
-    //   state.loading = true
-    // })
-    // builder.addCase(addNewWallet.rejected, state => {
-    //   state.loading = false
-    // })
-    // builder.addCase(addNewWallet.fulfilled, state => {
-    //   state.loading = false
-    // })
   },
 })
